pyproject.toml
[tool.poetry]
name = "ai4b-bugtracker"
version = "0.1.0"
description = "Production-ready bug reporting API"
authors = ["AI4B Candidate <candidate@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.2.1"
email-validator = "^2.1.1"
python-multipart = "^0.0.9"
redis = "^5.0.3"
slowapi = "^0.1.9"
rich = "^13.7.0"
httpx = "^0.27.0"
bleach = "^6.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
ruff = "^0.4.3"
mypy = "^1.8.0"
pytest-cov = "^5.0.0"
debugpy = "^1.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
